Fix nibegw protocol after first test runs
- little vs big endian - remove wrong byteswap
  - redefine nibe device addresses since they use big endian
- add missing length field to NibeRequest structure
- ignore known but unimplemented nibe commands
- improve logging

+++ main/nibegw_rs485.h
@@ -73,7 +73,7 @@ class NibeGw final : AbstractNibeGw {
     void sendData(const uint8_t* const data, uint8_t len);
     void sendAck();
     void sendNak();
-    bool shouldAckNakSend(uint8_t address);
+    bool shouldAckNakSend(uint16_t address);
 
     void connect();
     void disconnect();
